backend/requirements.txt
Flask==3.0.2
Flask-CORS==4.0.0
python-dotenv==1.0.1
supabase==1.2.0
PyJWT==2.8.0
requests==2.31.0
werkzeug==3.0.1